Codefactor refactoring
