+++ src/Osu.Music.UI/Utility/Proxies/DialogEditPlaylistProxy.cs
@@ -1,0 +1,8 @@
+﻿using Osu.Music.UI.ViewModels;
+
+namespace Osu.Music.UI.Utility.Proxies
+{
+    public sealed class DialogEditPlaylistProxy : BindingProxy<DialogEditPlaylistViewModel>
+    {
+    }
+}
